Relocate all components to /components
